+++ src/Pixeval/View/NavMenu.xaml.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace Pixeval.View
+{
+    /// <summary>
+    /// NavMenu.xaml 的交互逻辑
+    /// </summary>
+    public partial class NavMenu : UserControl
+    {
+        public NavMenu()
+        {
+            InitializeComponent();
+        }
+    }
+}
